requirements.txt
google-cloud-pubsub>=2.13.0
pandas>=1.5.0
pyarrow>=9.0.0
pydantic>=1.10.0
python-dotenv>=0.21.0 